Cleans up undeeded files and makes files generic to any case
