+++ src/Pendulum/case1/sketch.js
@@ -1,19 +1,90 @@
 'use strict';
 
-let EngineHandler = require('./EngineHandler.js');
-let RenderHandler = require('./RenderHandler.js');
-let State = require('./State.js');
+let State = require('../State.js');
+let Pendulum = require('../Pendulum.js');
 
+const CANVAS_WIDTH = 800;
+const CANVAS_HEIGHT = 600;
+
+let Engine = Matter.Engine;
 let Render = Matter.Render;
 let World = Matter.World;
+let Bodies = Matter.Bodies;
+let Constraint = Matter.Constraint;
 
-RenderHandler.setupCanvas();
+let engine = Engine.create();
 
-EngineHandler.createWorld(); // add bodies to canvas
+let render = Render.create({
+    element: document.getElementById('canvas'),
+    engine: engine,
+    options: {
+        width: CANVAS_WIDTH,
+        height: CANVAS_HEIGHT,
+        wireframes: false
+    }
+ });
 
-RenderHandler.startRender(); // allow for the rendering of frames of the world
+createWorld(); // add bodies to canvas
 
-RenderHandler.renderLoop(); // renders frames to the canvas
+Render.run(render); // allow for the rendering of frames of the world
+
+renderLoop(); // renders frames to the canvas
+
+/*
+  * Renders frames to send to the canvas
+*/
+
+function renderLoop() {
+  if (State.getIsPausedFlag()) { // the world is paused
+    requestAnimationFrame(renderLoop); // render next frame
+  }
+  else {
+    Engine.update(engine, 1000 / 60); // update at 60 FPS
+    requestAnimationFrame(renderLoop); // render next frame
+
+    Pendulum.setPendulumAngle(Pendulum.calculateAngle(Pendulum.string.bodies[0].position, Pendulum.pendulumWeight.position));
+    Pendulum.displayPendulumAngle();
+  }
+}
+
+/*
+  * Determines whether to continue rendering the world or not
+  * @param {boolean} isPaused - Flag to tell if the world is paused or not
+*/
+
+function determineRender(isPaused) {
+  if (isPaused) { // the world is paused
+    Render.stop(render);
+  }
+  else {
+    Render.run(render);
+  }
+}
+
+/*
+  * Sets up initial bodies of the world
+*/
+
+function createWorld() {
+  World.add(engine.world, [
+     Bodies.rectangle(400, 0, 800, 50, { isStatic: true }),
+     Bodies.rectangle(400, 600, 800, 50, { isStatic: true }),
+     Bodies.rectangle(800, 300, 50, 600, { isStatic: true }),
+     Bodies.rectangle(0, 300, 50, 600, { isStatic: true })
+  ]);
+
+  Pendulum.pendulumWeight = Bodies.circle(100, 100, 40, { mass: 0.04, frictionAir: 0});
+
+  let protractor = Bodies.circle(400, 50, 60, { isStatic: true});
+
+  World.add(engine.world, [Pendulum.pendulumWeight, protractor]);
+
+  Pendulum.string = World.add(engine.world, Constraint.create({
+    bodyA: protractor,
+    bodyB: Pendulum.pendulumWeight,
+    length: 0,
+  }));
+}
 
 /*
   * Pauses or unpauses the world from rendering
@@ -21,7 +92,7 @@ RenderHandler.renderLoop(); // renders frames to the canvas
 
 document.getElementById('pause-button').onclick = function() {
   State.setIsPausedFlag(State.getIsPausedFlag());
-  RenderHandler.determineRender(State.getIsPausedFlag());
+  determineRender(State.getIsPausedFlag());
 };
 
 /*
@@ -29,13 +100,11 @@ document.getElementById('pause-button').onclick = function() {
 */
 
 document.getElementById('reset-button').onclick = function() {
-  let engine = EngineHandler.getEngine(); // fetch a reference to the world engine
-
   World.clear(engine.world);
-  EngineHandler.createWorld();
+  createWorld();
 
   if (State.isPausedFlag === true) {
     State.setIsPausedFlag(State.getIsPausedFlag());
-    RenderHandler.startRender();
+    Render.start(render);
   }
 };
